Put Mal favorites in one table, put some indexes, misc fixes speed-ups

+++ src/PaperMalKing.MyAnimeList.Wrapper/Models/Favorites/FavoriteManga.cs
@@ -4,11 +4,11 @@ namespace PaperMalKing.MyAnimeList.Wrapper.Models.Favorites;
 
 internal sealed class FavoriteManga : BaseListFavorite
 {
-	internal FavoriteManga(string type, uint startYear, MalUrl url, string name, string? imageUrl) : base(type,
+	internal FavoriteManga(string type, ushort startYear, MalUrl url, string name, string? imageUrl) : base(type,
 		startYear, url, name, imageUrl)
 	{ }
 
-	internal FavoriteManga(string type, uint startYear, BaseFavorite baseFav) : base(type, startYear, baseFav)
+	internal FavoriteManga(string type, ushort startYear, BaseFavorite baseFav) : base(type, startYear, baseFav)
 	{ }
 
 	internal FavoriteManga(BaseListFavorite other) : base(other)
